packets now compute their header length
knowing the complete length of the packet did not allow to know the length of the contained payload nor the header. by separately computing one of them, the other may also be computed. Because, while constructing a packet, the payload length may be unknown, one may still want to compute the length of the already constructed header